Liten refactor
